Removed slider hitcircle following sliderball when it hasnt been hit, now it's stationary. Fixed aspire -1 duration sliders. Various fixes and shit. Removed GC.Collect when changing beatmaps. Forgot the rest of the changes! :D

+++ Easy2D/Modern/GLBuffer.cs
@@ -88,7 +88,7 @@ namespace Easy2D
             uint sizeInBytes = objectCount * sizeOfTypeInBytes;
             uint offsetInBytes = startIndex * sizeOfTypeInBytes;
 
-            GL.Instance.BufferSubData(bufferType, (IntPtr)offsetInBytes, (nuint)sizeInBytes, data);
+            GL.Instance.BufferSubData(bufferType, (IntPtr)offsetInBytes, sizeInBytes, data);
 
         }
